conexion-db

+++ src/main/java/com/techlab/crud/repository/Articulo/ArticuloRepository.java
@@ -1,7 +1,9 @@
 package com.techlab.crud.repository.Articulo;
 
-import com.techlab.crud.model.Articulo;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.techlab.crud.model.Articulo.Articulo;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -11,7 +13,8 @@ public interface ArticuloRepository extends JpaRepository<Articulo, Long> {
     List<Articulo> findByActivoFalse();
     List<Articulo> findByNombreContainingAndActivoTrue(String nombre);
     List<Articulo> findByCategoriaIdAndActivoTrue(Long categoriaId);
-    List<Articulo> findByPrecioLessThanEqualAndActivoTrue(int precio);
+    // Use Double to match Articulo.precio type and avoid implicit narrowing
+    List<Articulo> findByPrecioLessThanEqualAndActivoTrue(Double precio);
 
     Optional<Articulo> findByNombreAndMarca(String nombre, String marca);
 }
